new tests' structure
