+++ tests/test-runner.sh
@@ -1,0 +1,61 @@
+#!/bin/bash
+
+if [ $# -ne 1 ]; then
+  printf "Usage: %s SOURCEFILENAME" "$0"
+  exit 1
+fi
+
+command -v clang >/dev/null 2>&1 || { echo >&2 "clang is required but it's not installed. Aborting..."; exit 1; }
+command -v llvm-objcopy >/dev/null 2>&1 || { echo >&2 "llvm-objcopy is required but it's not installed. Aborting..."; exit 1; }
+command -v opt >/dev/null 2>&1 || { echo >&2 "opt is required but it's not installed. Aborting..."; exit 1; }
+
+remove_old_file() {
+    local FILENAME
+    FILENAME="$1"
+    if [ -e "${FILENAME}" ]; then
+        rm "${FILENAME}"
+        printf "Old %s file removed...\n" "${FILENAME}"
+    fi
+}
+
+SOURCEFILENAME="$1"
+BUILDPATH=$(realpath "$(dirname "$SOURCEFILENAME")/../build")
+BUILDTESTSPATH="$BUILDPATH/tests"
+SHAREDOBJECTFILE="$BUILDPATH/lib/libdaedalus.so"
+SOURCEFILEBASENAMEWEXT=$(basename "$SOURCEFILENAME" | sed 's/\.[^.]*$//')
+SOURCEFILENAMELL="$BUILDTESTSPATH/$SOURCEFILEBASENAMEWEXT.ll"
+SOURCEFILENAMEDLL="$BUILDTESTSPATH/$SOURCEFILEBASENAMEWEXT.d.ll"
+SLICESREPORTLOGFILE="$BUILDTESTSPATH/${SOURCEFILEBASENAMEWEXT}_slices_report.log"
+TRANSFORMATIONLOGFILE="$BUILDTESTSPATH/${SOURCEFILEBASENAMEWEXT}_transformation.log"
+ORIGINAL_EXECUTABLE="$BUILDTESTSPATH/$SOURCEFILEBASENAMEWEXT.bin"
+FINAL_EXECUTABLE="$BUILDTESTSPATH/$SOURCEFILEBASENAMEWEXT.d.bin"
+
+clang -Os -flto -fuse-ld=lld -Wl,--plugin-opt=-lto-embed-bitcode=post-merge-pre-opt "$SOURCEFILENAME" -o "$ORIGINAL_EXECUTABLE"
+llvm-objcopy --dump-section .llvmbc="$SOURCEFILENAMELL" "$ORIGINAL_EXECUTABLE"
+
+if [ -e "$SOURCEFILENAMELL" ]; then
+    opt -S -passes=mem2reg,lcssa "$SOURCEFILENAMELL" -o "$SOURCEFILENAMELL"
+
+    remove_old_file "$SLICESREPORTLOGFILE"
+    TESTLOGNAME="$TRANSFORMATIONLOGFILE"
+    remove_old_file "$TESTLOGNAME"
+
+    opt -stats -debug-only=Daedalus -passes=daedalus -load-pass-plugin="$SHAREDOBJECTFILE" -S "$SOURCEFILENAMELL" -o "$SOURCEFILENAMEDLL" &>> "$TESTLOGNAME"
+
+    if [ -e "$SOURCEFILENAMEDLL" ]; then
+        clang -Os "$SOURCEFILENAMEDLL" -o "$FINAL_EXECUTABLE"
+    fi
+fi
+
+if [ -e "$FINAL_EXECUTABLE" ]; then
+    "$FINAL_EXECUTABLE" 12 > "${SOURCEFILEBASENAMEWEXT}.output"
+fi
+if [ -e "$ORIGINAL_EXECUTABLE" ]; then
+    "$ORIGINAL_EXECUTABLE" 12 > "${SOURCEFILEBASENAMEWEXT}.reference_output"
+fi
+
+if cmp -s "${SOURCEFILEBASENAMEWEXT}.output" "${SOURCEFILEBASENAMEWEXT}.reference_output"; then
+    exit 0
+else
+    exit 1
+fi
