--- tests/test0.c
@@ -1,5 +1,0 @@
-#include <stdio.h>
-int main(void){
-	int a = 2;
-	printf("%d\n", a*3);
-}
