--- tests/test10.c
@@ -1,6 +1,0 @@
-#include <stdlib.h>
-int main(int argc, char **argv){
-	for(int i=0;i<atoi(argv[1]);++i){
-		i++;
-	}
-}
