--- tests/test15.c
@@ -1,6 +1,0 @@
-#include <stdio.h>
-
-int main(int argc, char **argv){
-    int a = argc*argc;
-    printf("%d", a*argc);
-}
