--- tests/test18.c
@@ -1,7 +1,0 @@
-#include <stdio.h>
-int main(int argc, char **argv){
-	int a = argc*3;
-	int b = argc*5;
-	printf("%d\n", a);
-	printf("%d\n", b);
-}
