--- tests/test22.c
@@ -1,74 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-
-struct Node {
-    int data;
-    struct Node* next;
-};
-
-void insertEnd(struct Node** head, int data) {
-    struct Node* newNode = (struct Node*)malloc(sizeof(struct Node));
-    newNode->data = data;
-    newNode->next = NULL;
-
-    if (*head == NULL) {
-        *head = newNode;
-        return;
-    }
-
-    struct Node* temp = *head;
-    while (temp->next != NULL) {
-        temp = temp->next;
-    }
-    temp->next = newNode;
-}
-
-void deleteValue(struct Node** head, int value) {
-    struct Node *temp = *head, *prev = NULL;
-
-    if (temp != NULL && temp->data == value) {
-        *head = temp->next;
-        free(temp);
-        return;
-    }
-
-    while (temp != NULL && temp->data != value) {
-        prev = temp;
-        temp = temp->next;
-    }
-
-    if (temp == NULL) return;
-
-    prev->next = temp->next;
-    free(temp);
-}
-
-void printList(struct Node* head) {
-    struct Node* temp = head;
-    while (temp != NULL) {
-        printf("%d -> ", temp->data);
-        temp = temp->next;
-    }
-    printf("NULL\n");
-}
-
-int main() {
-    struct Node* head = NULL;
-
-    insertEnd(&head, 1);
-    insertEnd(&head, 2);
-    insertEnd(&head, 3);
-    printList(head);
-
-    deleteValue(&head, 2);
-    printList(head);
-
-    deleteValue(&head, 1);
-    printList(head);
-
-    deleteValue(&head, 3);
-    printList(head);
-
-    return 0;
-}
-
