--- tests/test24.c
@@ -1,7 +1,0 @@
-#include <stdio.h>
-
-int main(int argc, char **argv){
-    int a = argc*argc;
-    int b = 2*a + 2;
-    int c = 3*a - 6;
-}
