--- tests/test33.c
@@ -1,9 +1,0 @@
-#include <stdio.h>
-
-static int a = 2;
-static int b = 2;
-
-int main(void){
-    printf("%d", a + b);
-    return a+b;
-}
