--- tests/test4.c
@@ -1,13 +1,0 @@
-#include <stdio.h>
-
-int g(){
-	return 2;
-}
-
-int f(){
-	return g();
-}
-
-int main(){
-	f();
-}
