--- tests/test8.c
@@ -1,9 +1,0 @@
-#include <stdio.h>
-
-int main(int argc, char **argv){
-
-	for(int i=argc;i<argc+3;++i){
-		printf("%d\n", i);
-	}
-
-}
