--- tests/test9.c
@@ -1,14 +1,0 @@
-#include <stdio.h>
-
-int main(int argc, char **argv){
-	int a2;
-	scanf("%d\n", &a2);
-	int a3;
-	scanf("%d\n", &a3);
-	if(a2) a3 = a2;
-	int a4 = a2;
-	printf("%d\n", a4);
-	printf("%d\n", a3);
-	
-}
-
